+++ crates/miden-standards/src/account/auth/mod.rs
@@ -7,8 +7,8 @@ pub use fee::{FeeConversionInfo, commit_fee_conversion_info};
 mod no_auth;
 pub use no_auth::NoAuth;
 
-mod pass_through;
-pub use pass_through::AuthPassThrough;
+mod tx_fee_collector;
+pub use tx_fee_collector::AuthTxFeeCollector;
 
 mod singlesig;
 pub use singlesig::AuthSingleSig;