+++ crates/miden-standards/src/account/components/mod.rs
@@ -9,8 +9,8 @@ use crate::account::auth::{
     AuthMultisig,
     AuthMultisigSmart,
     AuthNetworkAccount,
-    AuthPassThrough,
     AuthSingleSig,
+    AuthTxFeeCollector,
     NoAuth,
 };
 use crate::account::faucets::FungibleFaucet;
@@ -38,7 +38,7 @@ pub enum StandardAccountComponent {
     AuthGuardedMultisig,
     AuthNoAuth,
     AuthNetworkAccount,
-    AuthPassThrough,
+    AuthTxFeeCollector,
 }
 
 impl StandardAccountComponent {
@@ -59,7 +59,7 @@ impl StandardAccountComponent {
             Self::AuthGuardedMultisig => AuthGuardedMultisig::code(),
             Self::AuthNoAuth => NoAuth::code(),
             Self::AuthNetworkAccount => AuthNetworkAccount::code(),
-            Self::AuthPassThrough => AuthPassThrough::code(),
+            Self::AuthTxFeeCollector => AuthTxFeeCollector::code(),
         };
 
         code.procedure_roots()
@@ -121,8 +121,8 @@ impl StandardAccountComponent {
                 Self::AuthNetworkAccount => {
                     component_interface_vec.push(AccountComponentInterface::AuthNetworkAccount)
                 },
-                Self::AuthPassThrough => {
-                    component_interface_vec.push(AccountComponentInterface::AuthPassThrough)
+                Self::AuthTxFeeCollector => {
+                    component_interface_vec.push(AccountComponentInterface::AuthTxFeeCollector)
                 },
             }
         }
@@ -149,6 +149,6 @@ impl StandardAccountComponent {
         Self::AuthMultisigSmart.extract_component(procedures_set, component_interface_vec);
         Self::AuthNoAuth.extract_component(procedures_set, component_interface_vec);
         Self::AuthNetworkAccount.extract_component(procedures_set, component_interface_vec);
-        Self::AuthPassThrough.extract_component(procedures_set, component_interface_vec);
+        Self::AuthTxFeeCollector.extract_component(procedures_set, component_interface_vec);
     }
 }